Correctly handle Amazon GW forbidden errors
The Amazon API gateway used by the Easee API seems to route some requests to a blackhole, possibly because we keep trying to login with outdated credentials. However, such responses should result in a Forbidden error to make sure they're dealt with correctly upstream.

This PR adds a Faraday middleware that detects the Amazon Forbidden response header and changes the response status code to 403.